requirements.txt
opencv-python
torch
torchvision
Pillow
numpy
maxapi
aiohttp
python-dotenv
watchdog